@@ -8,12 +8,12 @@ class Finca {
         this.nombreTecnico,
     });
 
-    String id;
-    String nombreFinca;
-    String nombreProductor;
-    double areaFinca;
-    int tipoMedida;
-    String nombreTecnico;
+    String? id;
+    String? nombreFinca;
+    String? nombreProductor;
+    double? areaFinca;
+    int? tipoMedida;
+    String? nombreTecnico;
 
     factory Finca.fromJson(Map<String, dynamic> json) => Finca(
         id: json["id"],
