@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class DesicionesList extends StatelessWidget {
-    const DesicionesList({Key key}) : super(key: key);
+    const DesicionesList({Key? key}) : super(key: key);
 
     
     Future getRegistros() async{
@@ -17,11 +17,11 @@ class DesicionesList extends StatelessWidget {
         return listAcciones;
     }
 
-    Future getDatos(String id) async{
+    Future getDatos(String? id) async{
         
-        TestPiso testPiso = await DBProvider.db.getTestId(id);
-        Finca finca = await DBProvider.db.getFincaId(testPiso.idFinca);
-        Parcela parcela = await DBProvider.db.getParcelaId(testPiso.idLote);
+        TestPiso? testPiso = await (DBProvider.db.getTestId(id));
+        Finca? finca = await DBProvider.db.getFincaId(testPiso!.idFinca);
+        Parcela? parcela = await DBProvider.db.getParcelaId(testPiso.idLote);
 
         return [testPiso, finca, parcela];
     }
