@@ -1,15 +1,14 @@
 import 'dart:convert';
-
 import 'package:app_piso/src/models/acciones_model.dart';
 import 'package:app_piso/src/models/decisiones_model.dart';
 import 'package:app_piso/src/models/testPiso_model.dart';
-//import 'package:app_piso/src/pages/decisiones/pdf_view.dart';
 import 'package:app_piso/src/providers/db_provider.dart';
 import 'package:app_piso/src/models/selectValue.dart' as selectMap;
 import 'package:app_piso/src/utils/constants.dart';
 import 'package:app_piso/src/utils/widget/titulos.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_swiper/flutter_swiper.dart';
+import 'package:flutter_swiper_null_safety/flutter_swiper_null_safety.dart';
+
 
 class ReportePage extends StatefulWidget {
 
@@ -36,30 +35,30 @@ class _ReportePageState extends State<ReportePage> {
     
     
 
-    Future getdata(String idTest) async{
+    Future getdata(String? idTest) async{
 
         List<Decisiones> listDecisiones = await DBProvider.db.getDecisionesIdTest(idTest);         
         List<Acciones> listAcciones= await DBProvider.db.getAccionesIdTest(idTest);
-        TestPiso testplaga = await DBProvider.db.getTestId(idTest);
+        TestPiso? testplaga = await (DBProvider.db.getTestId(idTest));
 
-        Finca finca = await DBProvider.db.getFincaId(testplaga.idFinca);
-        Parcela parcela = await DBProvider.db.getParcelaId(testplaga.idLote);
+        Finca? finca = await DBProvider.db.getFincaId(testplaga!.idFinca);
+        Parcela? parcela = await DBProvider.db.getParcelaId(testplaga.idLote);
 
         return [listDecisiones, listAcciones, finca, parcela];
     }
 
     
     
-    Future<double> _countPercentTotal(String idTest,int idPlaga) async{
+    Future<double> _countPercentTotal(String? idTest,int idPlaga) async{
         double countPalga = await DBProvider.db.countPisoTotal(idTest, idPlaga);         
         return countPalga*100;
     }
 
-    Future<double> _countTotalCompetencia(String idTest,int idPlaga) async{
+    Future<double> _countTotalCompetencia(String? idTest,int idPlaga) async{
         double countPalga = await DBProvider.db.countCompetencia(idTest, idPlaga);        
         return countPalga*100;
     }
-    Future<double> _countTotalNoCompetencia(String idTest,int idPlaga) async{
+    Future<double> _countTotalNoCompetencia(String? idTest,int idPlaga) async{
         double countPalga = await DBProvider.db.countNoCompetencia(idTest, idPlaga);        
         return countPalga*100;
     }
@@ -69,7 +68,7 @@ class _ReportePageState extends State<ReportePage> {
 
     @override
     Widget build(BuildContext context) {
-        String idTest = ModalRoute.of(context).settings.arguments;
+        String? idTest = ModalRoute.of(context)!.settings.arguments as String?;
 
         return Scaffold(
             appBar: AppBar(),
@@ -79,7 +78,7 @@ class _ReportePageState extends State<ReportePage> {
                     if (!snapshot.hasData) {
                         return CircularProgressIndicator();
                     }
-                    List<Widget> pageItem = List<Widget>();
+                    List<Widget> pageItem = [];
                     Finca finca = snapshot.data[2];
                     Parcela parcela = snapshot.data[3];
 
@@ -107,7 +106,7 @@ class _ReportePageState extends State<ReportePage> {
                                                 "Deslice hacia la derecha para continuar con el reporte",
                                                 textAlign: TextAlign.center,
                                                 style: Theme.of(context).textTheme
-                                                    .headline5
+                                                    .headline5!
                                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 16)
                                             ),
                                         ),
@@ -146,7 +145,7 @@ class _ReportePageState extends State<ReportePage> {
         );
     }
 
-    Widget _principalData(String plagaid, BuildContext context, Finca finca, Parcela parcela){
+    Widget _principalData(String? plagaid, BuildContext context, Finca finca, Parcela parcela){
     
          return Container(
             decoration: BoxDecoration(
@@ -170,7 +169,7 @@ class _ReportePageState extends State<ReportePage> {
                                                     "Porcentaje de cobertura",
                                                     textAlign: TextAlign.center,
                                                     style: Theme.of(context).textTheme
-                                                        .headline5
+                                                        .headline5!
                                                         .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                                                 ),
                                             )
@@ -199,14 +198,14 @@ class _ReportePageState extends State<ReportePage> {
                                                             Expanded(
                                                                 child: Container(
                                                                     padding: EdgeInsets.symmetric(horizontal: 20.0),
-                                                                    child: Text('Estado de piso', textAlign: TextAlign.start, style: Theme.of(context).textTheme.headline6
+                                                                    child: Text('Estado de piso', textAlign: TextAlign.start, style: Theme.of(context).textTheme.headline6!
                                                                                             .copyWith(fontSize: 16, fontWeight: FontWeight.w600)),
                                                                 ),
                                                             ),
                                                             
                                                             Container(
                                                                 width: 100,
-                                                                child: Text('Cobertura', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                                                child: Text('Cobertura', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                                                         .copyWith(fontSize: 16, fontWeight: FontWeight.w600)),
                                                             ),
                                                         ],
@@ -228,8 +227,8 @@ class _ReportePageState extends State<ReportePage> {
             
     }
     Widget _dataFincas( BuildContext context, Finca finca, Parcela parcela ){
-        String labelMedidaFinca;
-        String labelvariedad;
+        String? labelMedidaFinca;
+        String? labelvariedad;
 
         final item = selectMap.dimenciones().firstWhere((e) => e['value'] == '${finca.tipoMedida}');
         labelMedidaFinca  = item['label'];
@@ -351,7 +350,7 @@ class _ReportePageState extends State<ReportePage> {
             children: [
                 Padding(
                     padding: EdgeInsets.symmetric(horizontal: 15),
-                    child: Text(titulo, textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                    child: Text(titulo, textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                           .copyWith(fontSize: 16, fontWeight: FontWeight.w600)),
                 ),
                 Divider()
@@ -360,11 +359,11 @@ class _ReportePageState extends State<ReportePage> {
     }
 
 
-    Widget _countPlagas(String idTest, int caminata){
-        List<Widget> lisItem = List<Widget>();
+    Widget _countPlagas(String? idTest, int caminata){
+        List<Widget> lisItem = [];
 
         for (var i = 0; i < itemEnContato.length; i++) {
-            String labelPlaga = itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "1","label": "No data"})['label'];
+            String? labelPlaga = itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "1","label": "No data"})['label'];
             int idplga = int.parse(itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "100","label": "No data"})['value']);
             
             
@@ -426,7 +425,7 @@ class _ReportePageState extends State<ReportePage> {
     }
     
 
-    Widget _countCompetencia(String idTest, int idplga, String labelPlaga){
+    Widget _countCompetencia(String? idTest, int idplga, String? labelPlaga){
         return Row(
             mainAxisAlignment: MainAxisAlignment.end,
             children: [
@@ -467,7 +466,7 @@ class _ReportePageState extends State<ReportePage> {
         );
     }
 
-    Widget _countNoCompetencia(String idTest, int idplga, String labelPlaga){
+    Widget _countNoCompetencia(String? idTest, int idplga, String? labelPlaga){
         return Row(
             mainAxisAlignment: MainAxisAlignment.end,
             children: [
@@ -508,7 +507,7 @@ class _ReportePageState extends State<ReportePage> {
         );
     }
 
-    Widget _sueloDesnudo(String idTest, int idplga, String labelPlaga){
+    Widget _sueloDesnudo(String? idTest, int idplga, String? labelPlaga){
         return Row(
             mainAxisAlignment: MainAxisAlignment.end,
             children: [
@@ -540,7 +539,7 @@ class _ReportePageState extends State<ReportePage> {
 
 
     Widget _hierbasProblematicas(List<Decisiones> decisionesList){
-        List<Widget> listPrincipales = List<Widget>();
+        List<Widget> listPrincipales = [];
 
         listPrincipales.add(
             Column(
@@ -552,7 +551,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Hierbas que consideran problematicas",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -567,7 +566,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 1) {
-                String label = hierbaProblematica.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label = hierbaProblematica.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -612,7 +611,7 @@ class _ReportePageState extends State<ReportePage> {
     }
 
     Widget _competeciaValoracion(List<Decisiones> decisionesList){
-        List<Widget> listPrincipales = List<Widget>();
+        List<Widget> listPrincipales = [];
 
         listPrincipales.add(
             Column(
@@ -624,7 +623,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Competencia entre hierbas y cacao",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -639,7 +638,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 2) {
-                String label= itemCompetencia.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label= itemCompetencia.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -669,7 +668,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Valoración de cobertura del piso",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -684,7 +683,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 3) {
-                String label= itemValoracion.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label= itemValoracion.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -727,7 +726,7 @@ class _ReportePageState extends State<ReportePage> {
     }
 
     Widget _observaciones(List<Decisiones> decisionesList){
-        List<Widget> listPrincipales = List<Widget>();
+        List<Widget> listPrincipales = [];
 
         listPrincipales.add(
             Column(
@@ -739,7 +738,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Observaciones de suelo",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -754,7 +753,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 4) {
-                String label= itemObsSuelo.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label= itemObsSuelo.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -784,7 +783,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Observaciones de sombra",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -799,7 +798,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 5) {
-                String label= itemObsSombra.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label= itemObsSombra.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -842,7 +841,7 @@ class _ReportePageState extends State<ReportePage> {
     }
 
     Widget _obsManejo(List<Decisiones> decisionesList){
-        List<Widget> listPrincipales = List<Widget>();
+        List<Widget> listPrincipales = [];
 
         listPrincipales.add(
             Column(
@@ -854,7 +853,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "Observaciones de manejo",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -869,7 +868,7 @@ class _ReportePageState extends State<ReportePage> {
         for (var item in decisionesList) {
 
             if (item.idPregunta == 6) {
-                String label= itemObsManejo.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? label= itemObsManejo.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
 
                 listPrincipales.add(
 
@@ -915,7 +914,7 @@ class _ReportePageState extends State<ReportePage> {
 
 
     Widget _accionesMeses(List<Acciones> listAcciones){
-        List<Widget> listPrincipales = List<Widget>();
+        List<Widget> listPrincipales = [];
 
         listPrincipales.add(
             Column(
@@ -927,7 +926,7 @@ class _ReportePageState extends State<ReportePage> {
                                 "¿Qué acciones vamos a realizar y cuando?",
                                 textAlign: TextAlign.center,
                                 style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.w600, fontSize: 18)
                             ),
                         )
@@ -941,14 +940,14 @@ class _ReportePageState extends State<ReportePage> {
         
         for (var item in listAcciones) {
 
-                List<String> meses = [];
-                String label= listSoluciones.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
-                List listaMeses = jsonDecode(item.repuesta);
+                List<String?> meses = [];
+                String? label= listSoluciones.firstWhere((e) => e['value'] == '${item.idItem}', orElse: () => {"value": "1","label": "No data"})['label'];
+                List listaMeses = jsonDecode(item.repuesta!);
                 if (listaMeses.length==0) {
                     meses.add('Sin aplicar');
                 }
                 for (var item in listaMeses) {
-                    String mes = _meses.firstWhere((e) => e['value'] == '$item', orElse: () => {"value": "1","label": "No data"})['label'];
+                    String? mes = _meses.firstWhere((e) => e['value'] == '$item', orElse: () => {"value": "1","label": "No data"})['label'];
                     
                     meses.add(mes);
                 }
@@ -959,7 +958,7 @@ class _ReportePageState extends State<ReportePage> {
                     ListTile(
                         title: Text('$label',
                             style: Theme.of(context).textTheme
-                                    .headline5
+                                    .headline5!
                                     .copyWith(fontWeight: FontWeight.bold, fontSize: 16)
                         ),
                         subtitle: Text(meses.join(","+" ")),
@@ -993,11 +992,11 @@ class _ReportePageState extends State<ReportePage> {
     }
 
 
-    Widget _plagasPDF(String idTest, int caminata){
-        List<Widget> lisItem = List<Widget>();
+    Widget _plagasPDF(String? idTest, int caminata){
+        List<Widget> lisItem = [];
 
         for (var i = 0; i < itemEnContato.length; i++) {
-            String labelPlaga = itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "1","label": "No data"})['label'];
+            String? labelPlaga = itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "1","label": "No data"})['label'];
             int idplga = int.parse(itemEnContato.firstWhere((e) => e['value'] == '$i', orElse: () => {"value": "100","label": "No data"})['value']);
             lisItem.add(
                 Row(
