@@ -35,10 +35,10 @@ class _AgregarTestState extends State<AgregarTest> {
     String _fecha = '';
     TextEditingController _inputfecha = new TextEditingController();
 
-    List<TestPiso> mainlistpisos ;
+    List<TestPiso>? mainlistpisos ;
 
-    List mainparcela;
-    TextEditingController _control;
+    List? mainparcela;
+    late TextEditingController _control;
 
     @mustCallSuper
     // ignore: must_call_super
@@ -82,7 +82,7 @@ class _AgregarTestState extends State<AgregarTest> {
                                         child:Text(
                                             'Método Punta de zapato',
                                             style: Theme.of(context).textTheme
-                                                .headline6
+                                                .headline6!
                                                 .copyWith(fontSize: 16)
                                         ),
                                     ),
@@ -97,7 +97,7 @@ class _AgregarTestState extends State<AgregarTest> {
                                                     child:Text(
                                                         '3 Caminatas',
                                                         style: Theme.of(context).textTheme
-                                                            .headline6
+                                                            .headline6!
                                                             .copyWith(fontSize: 16)
                                                     ),
                                                 ),
@@ -106,7 +106,7 @@ class _AgregarTestState extends State<AgregarTest> {
                                                     child:Text(
                                                         '20 Pasos por caminata',
                                                         style: Theme.of(context).textTheme
-                                                            .headline6
+                                                            .headline6!
                                                             .copyWith(fontSize: 16)
                                                     ),
                                                 ),
@@ -152,7 +152,7 @@ class _AgregarTestState extends State<AgregarTest> {
             items: _listitem,
             enabled: _enableFinca,
             validator: (value){
-                if(value.length < 1){
+                if(value!.length < 1){
                     return 'No se selecciono una finca';
                 }else{
                     return null;
@@ -186,9 +186,9 @@ class _AgregarTestState extends State<AgregarTest> {
                     controller: _control,
                     initialValue: '',
                     labelText: 'Seleccione la parcela',
-                    items: mainparcela,
+                    items: mainparcela as List<Map<String, dynamic>>,
                     validator: (value){
-                        if(value.length < 1){
+                        if(value!.length < 1){
                             return 'Selecione un elemento';
                         }else{
                             return null;
@@ -226,7 +226,7 @@ class _AgregarTestState extends State<AgregarTest> {
     }
 
     _selectDate(BuildContext context) async{
-        DateTime picked = await showDatePicker(
+        DateTime? picked = await showDatePicker(
             context: context,
 
             initialDate: new DateTime.now(),
@@ -262,7 +262,7 @@ class _AgregarTestState extends State<AgregarTest> {
 
                     label: Text('Guardar',
                         style: Theme.of(context).textTheme
-                            .headline6
+                            .headline6!
                             .copyWith(fontWeight: FontWeight.w600, color: Colors.white)
                     ),
                     padding:EdgeInsets.symmetric(vertical: 13, horizontal: 50),
@@ -284,13 +284,13 @@ class _AgregarTestState extends State<AgregarTest> {
 
         piso.caminatas = 3;
 
-        if  ( !formKey.currentState.validate() ){
+        if  ( !formKey.currentState!.validate() ){
             //Cuendo el form no es valido
             return null;
         }
-        formKey.currentState.save();
+        formKey.currentState!.save();
 
-        mainlistpisos.forEach((e) {
+        mainlistpisos!.forEach((e) {
             //print(piso.fechaTest);
             //print(e.fechaTest);
             if (piso.idFinca == e.idFinca && piso.idLote == e.idLote && piso.fechaTest == e.fechaTest) {
@@ -305,7 +305,7 @@ class _AgregarTestState extends State<AgregarTest> {
             return null;
         }
 
-        String checkParcela = mainparcela.firstWhere((e) => e['value'] == '${piso.idLote}', orElse: () => {"value": "1","label": "No data"})['value'];
+        String? checkParcela = mainparcela!.firstWhere((e) => e['value'] == '${piso.idLote}', orElse: () => {"value": "1","label": "No data"})['value'];
 
 
 
@@ -339,6 +339,6 @@ class _AgregarTestState extends State<AgregarTest> {
             duration: Duration(seconds: 2),
         );
 
-        scaffoldKey.currentState.showSnackBar(snackbar);
+        scaffoldKey.currentState!.showSnackBar(snackbar);
     }
 }
